:art: 修改為設計稿的5個Tabbar，改寫部分函數和跳轉。
原HomeScreen改為AllFuncScreen。
但HomeStack和Navigator沒看懂，於是沒改。
